template/pyproject.toml
[tool.poetry]
name = "exercise"
version = "0.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
streamlit = "^1.41.1"
matplotlib = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
